move resolvers into individual files
